Moving the code related to print string to the print_msg label
The implementation could use the stack for passing the parameters to the functions, for exemple, the messages in print_msg and the row value in set_cursor. Besides setting the parameter through registers is more efficient than using the stack, it's not possible manipulate the stack with less than 4 bytes.